Move Link and Overlay to components/common folder
