+++ client/src/components/common/Overlay/Overlay.module.scss.d.ts
