+++ client/src/components/index.ts
@@ -1,6 +1,4 @@
 export * from './AppBar';
 export * from './Layout';
-export * from './Link';
 export * from './MenuDrawer';
-export * from './Overlay';
 export * from './SearchBar';
